remove `from.rs`
