+++ src/lib.rs
@@ -2,5 +2,5 @@ pub mod de;
 pub mod error;
 pub mod value;
 
-pub use de::from::{from_file, from_path, from_read, from_str, from_str_raw};
+pub use de::{from_file, from_path, from_read, from_str, from_str_raw};
 pub use error::{JsonWithCommentError as Error, Result};
